Title: 2차원 최대 증가 수열, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 231220/2차원 최대 증가 수열/longest-increasing-sequence-2d.cpp
@@ -35,7 +35,7 @@ int main() {
       }
     }
   }
-  int result = 0;
+  short result = 0;
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < m; j++) {
       result = max(dp[i][j], result);
